mpw-manager-vue/backend/requirements.txt
Flask==2.3.3
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
Flask-Mail==0.9.1
openai==1.3.5
requests==2.31.0